// src/app/models/entities/Entidades.ts
export interface Adjunto {
  idAdjunto?: number;
  nombreArchivo?: string | undefined;
  extension?: string | undefined;
  mimeType?: string | undefined;
  tamaniobytes?: number | undefined;
  ruta?: string | undefined;
  producto?: Producto[] | undefined;
  usuario?: Usuario[] | undefined;
}

export interface AuthCredentials {
  username?: string | undefined;
  password?: string | undefined;
}

export interface DeleteAdjuntoCommand {
  idAdjunto?: number;
}

export interface DeleteHorariosCommand {
  idHorarios?: number;
}

export interface DeleteMaquinaCommand {
  idMaquinas?: number;
}

export interface DeleteUsuarioCommand {
  idUsuario?: number;
}

export interface Detallepedido {
  idDetallePedido?: number;
  idPedidos?: number;
  idProductos?: number;
  cantidad?: number | undefined;
  precioUnitario?: number | undefined;
  subtotal?: number | undefined;
  idPedidosNavigation?: Pedido;
  idProductosNavigation?: Producto;
}
// ...
export interface Pedido {
  idPedidos?: number;
  idUsuario?: number;
  fechaPedido?: Date;
  total?: number;
  esActivo?: boolean;
  detallepedido?: Detallepedido[] | undefined;
  factura?: Factura[] | undefined;
  idUsuarioNavigation?: Usuario;
}
// ...
export interface Producto {
  idProductos?: number;
  idAdjuntos?: number;
  nombreProdcuto?: string | undefined;
  precio?: number;
  stock?: number;
  esActivo?: boolean | undefined;
  detallepedido?: Detallepedido[] | undefined;
  idAdjuntosNavigation?: Adjunto;
}
// ...
export interface Usuario {
  idUsuario?: number;
  idAdjunto?: number | undefined;
  idPersonas?: number;
  idRol?: number | undefined;
  nombreUsuario?: string | undefined;
  clave?: string | undefined;
  esActivo?: boolean;
  factura?: Factura[] | undefined;
  idAdjuntoNavigation?: Adjunto;
  idPersonasNavigation?: Persona;
  idRolNavigation?: Rol;
  pedido?: Pedido[] | undefined;
  reserva?: Reserva[] | undefined;
}